Progress...

+++ src/main/java/cointoss/execution/ExecutionLog.java
@@ -12,7 +12,7 @@ package cointoss.execution;
 import static java.nio.charset.StandardCharsets.ISO_8859_1;
 import static java.nio.file.StandardOpenOption.APPEND;
 import static java.nio.file.StandardOpenOption.CREATE;
-import static java.util.concurrent.TimeUnit.SECONDS;
+import static psychopath.PsychopathOpenOption.ATOMIC_WRITE;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -155,7 +155,7 @@ public class ExecutionLog {
     };
 
     /** The log writer. */
-    private static final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1, run -> {
+    private static final ScheduledExecutorService scheduler = Executors.newSingleThreadScheduledExecutor(run -> {
         Thread thread = new Thread(run);
         thread.setName("ExecutionLog Writer");
         thread.setDaemon(true);
@@ -253,9 +253,11 @@ public class ExecutionLog {
         return I.signal(startDay)
                 .recurse(day -> day.plusDays(1))
                 .takeWhile(day -> day.isBefore(endDay) || day.isEqual(endDay))
-                .concatMap(day -> new Cache(day).read(type))
+                .concatMap(day -> {
+                    Cache cache = new Cache(day);
+                    return cache.read(type);
+                })
                 .effect(e -> cacheId = e.id)
-                .take(e -> e.isAfter(start))
                 .effectOnComplete(() -> storedId = cacheId)
                 .concat(network(-1).effect(this::cache));
     }
@@ -645,6 +647,10 @@ public class ExecutionLog {
             return this;
         }
 
+        long lastID() {
+            return 0;
+        }
+
         /**
          * Read cached date.
          * 
@@ -661,6 +667,7 @@ public class ExecutionLog {
                     return readCompact();
                 }
             } else if (existNormal()) {
+                repair();
                 return readNormal();
             } else {
                 ExecutionLogRepository external = service.externalRepository();
@@ -891,7 +898,7 @@ public class ExecutionLog {
             File compact = compactLog();
 
             try {
-                CsvWriter writer = buildCsvWriter(new ZstdOutputStream(compact.newOutputStream(), 1));
+                CsvWriter writer = buildCsvWriter(new ZstdOutputStream(compact.newOutputStream(ATOMIC_WRITE), 1));
 
                 return executions.maps(Market.BASE, (prev, e) -> {
                     writer.writeRow(logger.encode(prev, e));
@@ -915,7 +922,7 @@ public class ExecutionLog {
             File fast = fastLog();
 
             try {
-                CsvWriter writer = buildCsvWriter(new ZstdOutputStream(fast.newOutputStream(), 1));
+                CsvWriter writer = buildCsvWriter(new ZstdOutputStream(fast.newOutputStream(ATOMIC_WRITE), 1));
 
                 return executions.maps(Market.BASE, (prev, e) -> {
                     writer.writeRow(logger.encode(prev, e));
@@ -972,8 +979,8 @@ public class ExecutionLog {
          */
         void convertNormalToCompactAsync() {
             if (!existCompact() && (!queue.isEmpty() || normal.isPresent())) {
-                I.schedule(5, SECONDS).to(() -> {
-                    writeCompact(readNormal()).effectOnComplete(() -> normal.delete()).to(I.NoOP);
+                scheduler.execute(() -> {
+                    convertNormalToCompact();
                 });
             }
         }
@@ -1037,6 +1044,11 @@ public class ExecutionLog {
          * @return true if the compact log exists, false otherwise.
          */
         boolean repair() {
+            // ignore today
+            if (Chrono.utcToday().toLocalDate().isEqual(date)) {
+                return false;
+            }
+
             // confirm the completed compact log
             if (existCompact()) {
                 normal.delete();
@@ -1132,7 +1144,7 @@ public class ExecutionLog {
 
             if (completed) {
                 convertNormalToCompact();
-                return true;
+                return false;
             } else {
                 return false;
             }
